ff: Keep pinned tab icon aligned-center
When the tabs list is huge, the compact mode is enabled, tab browsing
left-right arrows appear.

In userchrome.css, we've reduced the pinned tab width, so we need to adjust
the tab content padding to keep the icon centered when the compact mode
is activate. If we don't, the default padding pushs the icon on the
right.

Signed-off-by: Thomas Venriès <thomas.venries@gmail.com>